requirements.txt
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.15.0
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
wandb>=0.12.0
tqdm>=4.61.0
tensorboard>=2.5.0
pillow>=8.2.0
pyyaml>=5.4.0
loguru>=0.5.0
pydicom>=2.2.0
# Medical imaging specific packages
monai>=1.0.0
# Optional for MRI-specific intensity normalization
intensity-normalization>=0.5.0